Formatting the sub-points in research-presentation.md
Fixed two errors: In point 5, there were two sub-points, but they were marked with "d" and "e", instead of "a" and "b". So I changed those pointers to be "a" and "b" instead.

Second Error was with the Markdown formatting:
When we use a,b,c,d to note down points, they are not automatically incremented like numbered lists, they'll be treated as regular text followed by a period.

Hence, adding an asterisk (*) mark before the a,b,c,d formats our page, so that it is easy to read like a numbered list, which was the intention in our original markdown.

I believe these two changes make the webpage easier to read, and also fixes the correctness -- changing "d" and "e" to "a" and "b".